buchner/project-template/requirements.txt
Fabric==1.6.0
Flask==0.9
Flask-Funnel==0.1.2
Flask-Mobility==0.1
Flask-Script==0.5.3
nose==1.2.1
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.7.2
WTForms==1.0.3

gunicorn==0.14.5
